Update and rename huffmancodetofile.cpp to binarycodetofile.cpp
Actually, this program can be used to put any binary codes into a file in a compact way. Like we have code length of 998bit than we will get a file of 125bytes. I still don't know how to get rid of the extra 2 bits here, but consider it insignificant when the code is long enough.